feat(dev): allow .env to be sourced when bringing up/down a cluster
Previously, cluster-up.sh and cluster-down.sh both had similar code to
clone the local-dev-cluster repo. This is now moved to a single script.

The scripts also `cd` in local-dev-cluster dir which meant that a .env
file at the root of kepler project will not be honored by the
local-dev-cluster/main.sh. This has been fixed by invoking the
main.sh from the root of kepler project itself

This commit modifies the cluster-up and down targets to
  * clone local-dev-cluster to tmp dir (which is ignored by git)
  * sources .env in project root where settings that needs to
    passed down to local-dev-cluster can be kept. E.g.
    ```sh
     PROMETHEUS_ENABLE=true
     GRAFANA_ENABLE=true
    ```
Signed-off-by: Sunil Thaha <sthaha@redhat.com>