fix(build): support newer version of kubectl
kubectl version --short has been deprecated since 1.23. In the latest
kubectl, the flag is removed and causes `build-manifest` to fail when
checking for the version.

This fixes the version check by using --client -o json options and using
jq to validate if the version is higher than expected - 1.21.

A new make target `tools` is added that installs all required project
tool to tmp/bin directory.

Signed-off-by: Sunil Thaha <sthaha@redhat.com>