chore(bpf): Fix Kepler Licensing
The eBPF code needs to have a GPL-compatible license.
This is due to its use of Linux Kernel structs (like task_struct)
and its use of GPL-only eBPF helpers (e.g bpf_probe_read) [1].

eBPF loaders default to GPL when loading code without a license
section. You can verify this by inspecting the loaded probes
with bpftool.

This commit adjusts the license of the eBPF files to be:

- GPL-2.0-only OR BSD-2-Clause for Kepler code
- GPL-2.0-only for the generated vmlinux.h files given they come from
  the Linux Kernel.

It adjusts the README to display these changes.

[1]: https://github.com/torvalds/linux/blob/c942a0cd3603e34dd2d7237e064d9318cb7f9654/kernel/trace/bpf_trace.c#L243-L250

Signed-off-by: Dave Tucker <dave@dtucker.co.uk>