fix(bpfassets): Fix object file lookup
For local development the instructions suggest running:

./_output/bin/linux_amd64/kepler

This checks for the bytecode in /var/lib/kepler/bpfassets
However, for local development this directory doesn't exist.

The fallback was to look in ../../../bpfassets/libbpf/bpf.o
Running the recommended command was causing kepler to look in
strange locations (i.e /bpfassets) for bytecode.

This PR fixes the lookup for local development to use a glob
pattern starting at the current directory. This works well
for local development, with the added bonus of making it easier
to test kepler builds on remote systems since you can also
copy the binary and bytecode files together, without having
to place the bytecode in a special path.

Signed-off-by: Dave Tucker <dave@dtucker.co.uk>