chore: Use Codecov
This removes the current infra for updating the code coverage badge
which calculates a percentage and checks in a commit to update the
README.

It's replaced with CodeCov, a web service that will provide an
always-up-to-date link to a coverage badge, as well as allowing
users to dig into coverage reports through a web UI.

This avoids flooding the git history with commits that update
the coverage badge. For example, the SHA of the latest git
commit and the SHA used to build the `latest` container image
aren't the same because of the coverage commit.
Similary, we miss seeing information about the last merge
commit (and a CI status indicator) on the front page on GitHub.

Signed-off-by: Dave Tucker <dave@dtucker.co.uk>