chore: Improve kepler_exporter_build_info metric
This adds:
- Git Revision
- Git Branch
- OS
- Arch

To do so, we drop the prometheus/common/version collector - which is
simply a wrapper around GuageFunc and replace it with our own.

Given this build info is specific to the exporter it doesn't make sense
to live in pkg/version. These variables are instead placed in
cmd/exporter (package main) and can be more easily overriden at compile
time by simply patching main.Version.

Signed-off-by: Dave Tucker <dave@dtucker.co.uk>