ci: Don't override binary version in image builds
Before this change, the `kepler_exporter_build_info` metric would report
that the current version is `latest`. This makes it difficult to
understand precisely which version of Kepler is running.

If we don't override this at image build time, then the version will
be reported as something like v0.7.10-91-g95dce23a

This more accurately identifies which version of kepler is in use.

Signed-off-by: Dave Tucker <dave@dtucker.co.uk>