fix(bpf): Use BTF-Defined Raw Tracepoints
It turns out that RHEL is carrying a patch that increases the size of
the unreadable tracepoint-common area from 8 bytes to 12 bytes.
This means that our reads into the context area are off, causing
the probes to return invalid data.

To fix this we use BTF-Defined Raw Tracepoints that were added in
Kernel 5.5 (less than our 5.12 minimum version). This type of
tracepoint program uses BTF and CO-RE to access arguments
from the tracepoint context.

Signed-off-by: Dave Tucker <dave@dtucker.co.uk>