fix(bpf): Access __state from task_struct
In older kernels prev_state isn't passed as part of the context of the
function. Therefore it's better to access prev_task-->_state directly.

The field was renamed from state to __state in Kernel 5.14.
Given that we need to support Kernel 5.12+ we include the
bpf_core_read.h file such that we can use the bpf_core_field_exists
function that allows us to see which field is present in task struct
and return the correct value.

Signed-off-by: Dave Tucker <dave@dtucker.co.uk>