chore: Flatten Uint64Stat
This commit fixes a couple of small issues.
Firstly, in Go maps are reference types (like pointers) so
using *map[string]any for example isn't required.
Secondly, this removes the structure that backs
Uint64StatCollection and replaces it with a map.

Finally, operations in Uint64Stat have been made
atomic to ensure correct operation if methods are
called from different threads. A test has been added
to verify this behaviour - this test fails when
run against the main branch.

Signed-off-by: Dave Tucker <dave@dtucker.co.uk>