feat: customize vm_id with libvirt metadata
By default the name chosen to define the Virtual Machine ID is the one
found in /proc/<pid>/cgroup/.

Many IaaS platform like Openstack use metadata in libvirt to customise
among other things the instance's ID.

Allow the user to use the metadata in libvirt to set the name of the
vm metrics and gives a better user experience.

Set "LIBVIRT_METADATA_URI" with the uri of the XML namespace identifier.
Set "LIBVIRT_METADATA_TOKEN" to choose the string used for VM ID. By
default, the token "name" is used.

Signed-off-by: Anthony Harivel <aharivel@redhat.com>