feat(disable-power-models): Add DisablePowerModels flag
Added a config flag to prevent Kepler from resorting to models when
power meters like acpi and rapl are not available.

Issues: While Node level metrics that rely on models are fully removed,
due to the current setup of process metrics, process metrics which rely
on node metrics that rely on models are not removed and instead output 0
in prometheus.

Signed-off-by: Kaiyi <kaiyiliu21@gmail.com>