refactor: allow services to be initialized serially
This commit modifies the start up behavior. It refactors the service lifecycle
management, replacing the Start and Stop methods with
* Init: used to initiliaze a service and can assume it isn't called from
  go routines
* Run: Runs the service and blocks until context is cancelled or an
  error is encountered
* `Shutdown`:  methods is the cleanup and indicates that you can't call
  `Run` after `Shutdown`.

This change aims to separate initialization which is run serially and allows
services to initialize without worrying about thread-safety.
I.E. Its okay for an Init() to call another Init().

Thus this change separates initialization, runtime, and cleanup phases for better
clarity and control.

Tests are updated keep up with the change.

Signed-off-by: Sunil Thaha <sthaha@redhat.com>