chore(compose): update dashboard to support recent zone changes
With the recent zone changes to include zone+index in the name the
current dashboard fails to render the queries that don't follow the new
zone name format. This commit updates the dashboard to include a
`zone_clean` variable that filters out the zone index and then use that
variable where zone+index is not supported.

Signed-off-by: vprashar2929 <vibhu.sharma2929@gmail.com>