feat(rapl): support zone aggregation and topology energy calculation
This commit adds support for aggregating zones by name using a synthetic
AggregatedZone which implements EnergyZone and provides Energy that sums
up all the zones in the AggregatedZone.

CpuPowerMeter interface now has a new function TopologyEnergy() which is
the energy of the highest topology that is computed using TopologyEnergy
utility struct.

Signed-off-by: Sunil Thaha <sthaha@redhat.com>