chore(compose): add support for deploying redfish using sushy tool
This commit adds support for deploying sushy-tool using dev compose.
Sushy-tool provides a static redfish responder that uses mockups
enabling redfish API for testing and development.

The kepler-dev service now connects to redfish API endpoint to provide
platform metrics.

*NOTE: The mockups contain static Power data and are not real data.
This is for testing purposes only and not recommended for production.*

Signed-off-by: vprashar2929 <vibhu.sharma2929@gmail.com>